requirements.txt
asgiref==3.5.1
dj-database-url==0.5.0
Django==4.0.4
django-formtools==2.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.0.0
